### 2024-06-05 #84 @Harshvardhan-Mestha
[Re] Teaching CLIP to Count to Ten
**Original article:** Teaching CLIP to Count to Ten (https://arxiv.org/pdf/2302.12066)

**Scientific domain:**  Computer Vision
