+++ resc/metadata.yaml
@@ -15,7 +15,7 @@ authors:
     email: harshvardhanmestha@yahoo.com
     affiliations: 1,2,*
 
-  - name: Tejas Agarwal
+  - name: Tejas Agrawal
     email: tejasagrawal55@gmail.com
     affiliations: 1,2
 
