### 2024-10-11 #84 @rougier
@anthony-strock @ilemhadri @bengioewould @hhihnyou be interested in reviewing this submission:
Teaching CLIP to Count to Ten (https://arxiv.org/pdf/2302.12066)
