### 2025-04-11 #84 @a-strock
@rougier Happy to review this submission.
